requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
pymysql==1.0.2
Werkzeug~=3.0.3
python-dotenv~=1.0.1
pytest~=8.2.2
alembic~=1.13.1
SQLAlchemy~=2.0.30